+++ Cargo.toml
@@ -11,6 +11,7 @@ members = [
     "bin/rebaser",
     "bin/sdf",
     "bin/veritech",
+    "lib/naxum-api-types",
     "lib/auth-api-client",
     "lib/billing-events",
     "lib/buck2-resources",
@@ -32,6 +33,7 @@ members = [
     "lib/nats-subscriber",
     "lib/naxum",
     "lib/object-tree",
+    "lib/pending-events",
     "lib/pinga-core",
     "lib/pinga-server",
     "lib/permissions",
