+++ lib/naxum-api-types/BUCK
@@ -1,14 +1,17 @@
 load("@prelude-si//:macros.bzl", "rust_library")
 
 rust_library(
-    name = "rebaser-core",
+    name = "naxum-api-types",
     deps = [
-        "//lib/naxum-api-types:naxum-api-types",
         "//lib/si-data-nats:si-data-nats",
-        "//lib/si-events-rs:si-events",
+        "//third-party/rust:ciborium",
+        "//third-party/rust:derive_more",
         "//third-party/rust:remain",
         "//third-party/rust:serde",
+        "//third-party/rust:serde_json",
         "//third-party/rust:strum",
+        "//third-party/rust:thiserror",
+        "//third-party/rust:ulid",
     ],
     srcs = glob([
         "src/**/*.rs",
