+++ lib/naxum-api-types/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "rebaser-core"
+name = "naxum-api-types"
 version.workspace = true
 authors.workspace = true
 license.workspace = true
@@ -9,10 +9,13 @@ rust-version.workspace = true
 publish.workspace = true
 
 [dependencies]
-naxum-api-types = { path = "../../lib/naxum-api-types" }
 si-data-nats = { path = "../../lib/si-data-nats" }
-si-events = { path = "../../lib/si-events-rs" }
 
+ciborium = { workspace = true }
+derive_more = { workspace = true }
 remain = { workspace = true }
 serde = { workspace = true }
+serde_json = { workspace = true }
 strum = { workspace = true }
+thiserror = { workspace = true }
+ulid = { workspace = true }
