+++ lib/pending-events/BUCK
@@ -1,14 +1,16 @@
 load("@prelude-si//:macros.bzl", "rust_library")
 
 rust_library(
-    name = "rebaser-core",
+    name = "pending-events",
     deps = [
-        "//lib/naxum-api-types:naxum-api-types",
         "//lib/si-data-nats:si-data-nats",
         "//lib/si-events-rs:si-events",
+        "//lib/telemetry-nats-rs:telemetry-nats",
+        "//lib/telemetry-rs:telemetry",
         "//third-party/rust:remain",
         "//third-party/rust:serde",
-        "//third-party/rust:strum",
+        "//third-party/rust:serde_json",
+        "//third-party/rust:thiserror",
     ],
     srcs = glob([
         "src/**/*.rs",
