+++ lib/pending-events/Cargo.toml
@@ -1,18 +1,20 @@
 [package]
-name = "rebaser-core"
+name = "pending-events"
+edition = "2021"
 version.workspace = true
 authors.workspace = true
 license.workspace = true
 repository.workspace = true
-edition.workspace = true
 rust-version.workspace = true
 publish.workspace = true
 
 [dependencies]
-naxum-api-types = { path = "../../lib/naxum-api-types" }
 si-data-nats = { path = "../../lib/si-data-nats" }
 si-events = { path = "../../lib/si-events-rs" }
+telemetry = { path = "../../lib/telemetry-rs" }
+telemetry-nats = { path = "../../lib/telemetry-nats-rs" }
 
 remain = { workspace = true }
 serde = { workspace = true }
-strum = { workspace = true }
+serde_json = { workspace = true }
+thiserror = { workspace = true }
