+++ lib/rebaser-client/src/lib.rs
@@ -16,7 +16,9 @@ use si_data_nats::{
     jetstream::{self, Context},
     HeaderMap, Message, NatsClient, Subject,
 };
-use si_events::{rebase_batch_address::RebaseBatchAddress, ChangeSetId, WorkspacePk};
+use si_events::{
+    rebase_batch_address::RebaseBatchAddress, ChangeSetId, EventSessionId, WorkspacePk,
+};
 use telemetry::prelude::*;
 use telemetry_nats::propagation;
 use thiserror::Error;
@@ -94,9 +96,17 @@ impl Client {
         workspace_id: WorkspacePk,
         change_set_id: ChangeSetId,
         updates_address: RebaseBatchAddress,
+        event_session_id: EventSessionId,
     ) -> Result<RequestId> {
-        self.call_async(workspace_id, change_set_id, updates_address, None, None)
-            .await
+        self.call_async(
+            workspace_id,
+            change_set_id,
+            updates_address,
+            None,
+            None,
+            event_session_id,
+        )
+        .await
     }
 
     /// Asynchronously enqueues graph updates that originate from a Change Set & return a
@@ -116,6 +126,7 @@ impl Client {
         change_set_id: ChangeSetId,
         updates_address: RebaseBatchAddress,
         from_change_set_id: ChangeSetId,
+        event_session_id: EventSessionId,
     ) -> Result<RequestId> {
         self.call_async(
             workspace_id,
@@ -123,6 +134,7 @@ impl Client {
             updates_address,
             Some(from_change_set_id),
             None,
+            event_session_id,
         )
         .await
     }
@@ -143,12 +155,19 @@ impl Client {
         workspace_id: WorkspacePk,
         change_set_id: ChangeSetId,
         updates_address: RebaseBatchAddress,
+        event_session_id: EventSessionId,
     ) -> Result<(
         RequestId,
         BoxFuture<'static, Result<EnqueueUpdatesResponse>>,
     )> {
-        self.call_with_reply(workspace_id, change_set_id, updates_address, None)
-            .await
+        self.call_with_reply(
+            workspace_id,
+            change_set_id,
+            updates_address,
+            None,
+            event_session_id,
+        )
+        .await
     }
 
     /// Enqueues graph updates that originate from a Change Set and return a [`Future`] that will
@@ -168,6 +187,7 @@ impl Client {
         change_set_id: ChangeSetId,
         updates_address: RebaseBatchAddress,
         from_change_set_id: ChangeSetId,
+        event_session_id: EventSessionId,
     ) -> Result<(
         RequestId,
         BoxFuture<'static, Result<EnqueueUpdatesResponse>>,
@@ -177,6 +197,7 @@ impl Client {
             change_set_id,
             updates_address,
             Some(from_change_set_id),
+            event_session_id,
         )
         .await
     }
@@ -188,6 +209,7 @@ impl Client {
         updates_address: RebaseBatchAddress,
         from_change_set_id: Option<ChangeSetId>,
         maybe_reply_inbox: Option<&Subject>,
+        event_session_id: EventSessionId,
     ) -> Result<RequestId> {
         let id = RequestId::new();
 
@@ -197,6 +219,7 @@ impl Client {
             change_set_id,
             updates_address,
             from_change_set_id,
+            event_session_id: Some(event_session_id),
         });
 
         // Cut down on the amount of `String` allocations dealing with ids
@@ -245,6 +268,7 @@ impl Client {
         change_set_id: ChangeSetId,
         updates_address: RebaseBatchAddress,
         from_change_set_id: Option<ChangeSetId>,
+        event_session_id: EventSessionId,
     ) -> Result<(
         RequestId,
         BoxFuture<'static, Result<EnqueueUpdatesResponse>>,
@@ -272,6 +296,7 @@ impl Client {
                 updates_address,
                 from_change_set_id,
                 Some(&reply_inbox),
+                event_session_id,
             )
             .await?;
