+++ lib/rebaser-core/src/lib.rs
@@ -1,10 +1,8 @@
+//! This crate contains common information for the rebaser for clients, servers and interested parties.
+
 pub mod api_types;
-pub mod content_info;
 pub mod nats;
 
-pub use self::{
-    api_types::{
-        ApiVersionsWrapper, ApiWrapper, DeserializeError, RequestId, SerializeError, UpgradeError,
-    },
-    content_info::ContentInfo,
-};
+// TODO(nick): we should only export this within the "api_types" module. Delete this and fix the
+// compilation errors.
+pub use naxum_api_types::*;
