+++ lib/si-events-rs/src/event_session.rs
@@ -1,0 +1,3 @@
+use crate::id;
+
+id!(EventSessionId);
